+++ boot/boot.go
@@ -5,14 +5,17 @@ package boot
 import (
 	"flag"
 	"fmt"
+	"os"
 	"strings"
 	"time"
 
-	"github.com/goyek/goyek/v2"
-	"github.com/goyek/goyek/v2/middleware"
+	"github.com/goyek/goyek/v3"
+	"github.com/goyek/goyek/v3/middleware"
 
 	"github.com/goyek/x/color"
 )
+
+const exitCodeInvalid = 2
 
 // Reusable flags used by the build pipeline.
 var (
@@ -24,12 +27,25 @@ var (
 	noColor = flag.Bool("no-color", false, "disable colorizing output")
 )
 
-// Main is an extension of goyek.Main which additionally
-// defines flags and uses the most useful middlewares.
+// Main is an extension of goyek.Main which additionally defines reusable flags
+// and configures the flow with the most useful middlewares.
+//
+// It automatically sets up:
+//   - Colored output via [color.ReportFlow] and [color.ReportStatus]
+//   - Standard middlewares ([middleware.BufferParallel], [middleware.SilentNonFailed], [middleware.ReportLongRun])
+//   - Command line flags for common options (-v, -dry-run, -long-run, etc.)
+//
+// The command line syntax is: [tasks] [flags] [--] [args]
+//
+// To add custom flags, define them before calling Main (see the example).
 func Main() {
+	tasks, args := goyek.SplitTasks(os.Args[1:])
 	flag.CommandLine.SetOutput(goyek.Output())
 	flag.Usage = usage
-	flag.Parse()
+	if err := flag.CommandLine.Parse(args); err != nil {
+		fmt.Fprintln(goyek.Output(), err)
+		os.Exit(exitCodeInvalid)
+	}
 
 	if *dryRun {
 		*v = true // needed to report the task status
@@ -64,11 +80,11 @@ func Main() {
 
 	goyek.SetUsage(usage)
 	goyek.SetLogger(&color.CodeLineLogger{})
-	goyek.Main(flag.Args(), opts...)
+	goyek.Main(tasks, opts...)
 }
 
 func usage() {
-	fmt.Println("Usage of build: [flags] [--] [tasks]")
+	fmt.Println("Usage of build: [tasks] [flags] [--] [args]")
 	goyek.Print()
 	fmt.Println("Flags:")
 	flag.PrintDefaults()