+++ color/reportflow.go
@@ -4,7 +4,7 @@ import (
 	"time"
 
 	"github.com/fatih/color"
-	"github.com/goyek/goyek/v2"
+	"github.com/goyek/goyek/v3"
 )
 
 // ReportFlow is a middleware which reports the flow execution status with colors.