+++ CHANGELOG.md
@@ -6,6 +6,19 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this library adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html)
 as well as to [Module version numbering](https://go.dev/doc/modules/version-numbers).
 
+## [0.4.0](https://github.com/goyek/x/releases/tag/v0.4.0) - 2025-11-25
+
+### Changed
+
+- **BREAKING**: Change `boot.Main` command line syntax from `[flags] [--] [tasks]`
+  to `[tasks] [flags] [--] [args]` to align with `goyek` v3 recommendation.
+- Bump `github.com/goyek/goyek` to `3.0.0`.
+- Bump other dependencies.
+
+### Remove
+
+- Drop support for Go 1.22.
+
 ## [0.3.0](https://github.com/goyek/x/releases/tag/v0.3.0) - 2025-03-11
 
 ### Changed